// Iserv.Niis.Angular/src/app/requests/enums/status-codes.enum.ts
export enum StatusCodes {
  /* Успешно */
  Successfully = 1,
  /* Документ не найден */
  NotFound = 2,
  /* Тип заявки не совпадает */
  UnknownType = 3,
  /* Некорректные входные данные */
  BadRequest = 4,
  /* Документ уже был зарегистрирован */
  ExistNumber = 5
}